Filter native retweets and 0 retweet counts
The app needs to filter out native retweets and retweet counts of 0 in
order to work properly. Before, we were relying on instructions to the
user to setup their search text so those are filtered out at the Twitter
Search API layer. However, to be more robust, we should still do that
check in the app's code, but advise the user to filter at the Twitter
search API layer as an optimization and cost savings measure.